+++ parsec/data_dist/matrix/two_dim_rectangle_cyclic_band.c
@@ -20,10 +20,10 @@ static uint32_t twoDBC_band_rank_of(parsec_data_collection_t * desc, ...)
     va_end(ap);
 
     /* Check tile location within band_size */
-    if( (unsigned int)abs((int)m-n) < dc->band_size ){
+    if( (unsigned int)abs((int)m - (int)n) < dc->band_size ){
         /* The new m in band
          * (int)m - n + dc->band_size - 1 will not be negative in this scenario */
-        m = (unsigned int)((int)m - n + dc->band_size - 1);
+        m = (unsigned int)((int)m - (int)n + dc->band_size - 1);
         return dc->band.super.super.rank_of(&dc->band.super.super, m, n);
     }
 
@@ -44,10 +44,10 @@ static int32_t twoDBC_band_vpid_of(parsec_data_collection_t * desc, ...)
     va_end(ap);
 
     /* Check tile location within band_size */
-    if( (unsigned int)abs((int)m-n) < dc->band_size ){
+    if( (unsigned int)abs((int)m - (int)n) < dc->band_size ){
         /* The new m in band
          * (int)m - n + dc->band_size - 1 will not be negative in this scenario */
-        m = (unsigned int)((int)m - n + dc->band_size - 1);
+        m = (unsigned int)((int)m - (int)n + dc->band_size - 1);
         return dc->band.super.super.vpid_of(&dc->band.super.super, m, n);
     }
 
@@ -73,10 +73,10 @@ static parsec_data_t* twoDBC_band_data_of(parsec_data_collection_t *desc, ...)
 #endif
 
     /* Check tile location within band_size */
-    if( (unsigned int)abs((int)m-n) < dc->band_size ) {
+    if( (unsigned int)abs((int)m - (int)n) < dc->band_size ) {
         /* The new m in band
          * (int)m - n + dc->band_size - 1 will not be negative in this scenario */
-        m = (unsigned int)((int)m - n + dc->band_size - 1);
+        m = (unsigned int)((int)m - (int)n + dc->band_size - 1);
         return dc->band.super.super.data_of(&dc->band.super.super, m, n);
     }
